Rename missing Events
